+++ tags_test.go
@@ -7,13 +7,13 @@ import (
 func TestCreateTags(t *testing.T) {
 	req := &CreateTags{}
 	_ = req.Response().(*AsyncJobResult)
-	_ = req.AsyncResponse().(*booleanResponse)
+	_ = req.AsyncResponse().(*BooleanResponse)
 }
 
 func TestDeleteTags(t *testing.T) {
 	req := &DeleteTags{}
 	_ = req.Response().(*AsyncJobResult)
-	_ = req.AsyncResponse().(*booleanResponse)
+	_ = req.AsyncResponse().(*BooleanResponse)
 }
 
 func TestListTags(t *testing.T) {